--- src/lib/normalize-config.ts
@@ -1,15 +1,0 @@
-import { ExtensionConfiguration } from './vscode-service';
-
-export interface VSCodeStylableExtensionConfig {
-    diagnostics: {
-        ignore: string[];
-    };
-    formatting: {
-        newLineBetweenSelectors: boolean;
-        newLineBetweenRulesets: boolean;
-        endOfLine: string;
-        endWithNewLine: boolean;
-        experimental: boolean;
-        wrapLineLength: number;
-    };
-}
